Refeactor gitlab paginator
